#134 - Moved code to more logic location
