+++ source/asdmanager.py
@@ -22,7 +22,6 @@ Module for ASD Manager SetupController
 import os
 import sys
 import json
-import time
 import logging
 from source.tools.configuration.configuration import Configuration
 from source.tools.interactive import Interactive
@@ -30,7 +29,7 @@ from source.tools.toolbox import Toolbox
 from source.tools.services.service import ServiceManager
 from source.tools.localclient import LocalClient
 from source.tools.log_handler import LogHandler
-from subprocess import CalledProcessError, check_output
+from subprocess import check_output
 
 BOOTSTRAP_FILE = '/opt/asd-manager/config/bootstrap.json'
 PRECONFIG_FILE = '/opt/asd-manager/config/preconfig.json'
@@ -254,68 +253,6 @@ def remove(silent=None):
         local_client.file_delete(filenames=ArakoonConfiguration.CACC_LOCATION)
     local_client.file_delete(filenames=BOOTSTRAP_FILE)
     print '\n' + Interactive.boxed_message(['ASD Manager removal completed'])
-
-
-def monitor_services():
-    """
-    Monitor the ASD services
-    :return: None
-    """
-    try:
-        previous_output = None
-        service_manager = check_output('cat /proc/1/comm', shell=True).strip()
-        while True:
-            # Gather service states
-            running_services = {}
-            non_running_services = {}
-            longest_service_name = 0
-            if service_manager == 'systemd':
-                for service_name in check_output('systemctl list-unit-files --type=service | egrep "alba-|asd-" | tr -s " " | cut -d " " -f 1', shell=True).splitlines():
-                    try:
-                        service_state = check_output('systemctl is-active {0}'.format(service_name), shell=True).strip()
-                    except CalledProcessError as cpe:
-                        service_state = cpe.output
-
-                    if service_state == 'active':
-                        running_services[service_name] = service_state
-                    else:
-                        non_running_services[service_name] = service_state
-
-                    if len(service_name) > longest_service_name:
-                        longest_service_name = len(service_name)
-            else:
-                for service_info in check_output('initctl list | egrep "alba-|asd-"', shell=True).splitlines():
-                    service_info = service_info.split(',')[0].strip()
-                    service_name = service_info.split()[0].strip()
-                    service_state = service_info.split()[1].strip()
-                    if service_state == "start/running":
-                        running_services[service_name] = service_state
-                    else:
-                        non_running_services[service_name] = service_state
-
-                    if len(service_name) > longest_service_name:
-                        longest_service_name = len(service_name)
-
-            # Put service states in list
-            output = ['ASD Manager running processes',
-                      '=============================\n']
-            for service_name in sorted(running_services):
-                output.append('{0} {1} {2}'.format(service_name, ' ' * (longest_service_name - len(service_name)), running_services[service_name]))
-
-            output.extend(['\n\nASD Manager non-running processes',
-                           '=================================\n'])
-            for service_name in sorted(non_running_services):
-                output.append('{0} {1} {2}'.format(service_name, ' ' * (longest_service_name - len(service_name)), non_running_services[service_name]))
-
-            # Print service states (only if changes)
-            if previous_output != output:
-                print '\x1b[2J\x1b[H'
-                for line in output:
-                    print line
-                previous_output = list(output)
-            time.sleep(1)
-    except KeyboardInterrupt:
-        pass
 
 
 def _validate_and_retrieve_pre_config():
